Switch from boost::program_options to Lean Mean C++ Option Parser.
This change eliminates the need to link to boost, which means
that ncron will not break if a dynamically linked boost library
is updated to a new version without recompiling ncron.

Currently it also means that ncron will not have the ability
to parse a configuration file for command line options, but
I can add that back if needed.