Update the history file atomically.
I'm not using fsync() but I am using the standard strategy of
writing to a temporary and using rename() to overwrite the
old file with the updated file contained in the temporary.

The sole reason I'm not calling fsync() is that I don't want
to potentially write out to SSD and burn write cycles without
merit by default.