Allow more precise time constraints and optimize.
This changes the configuration file format for time constraints.
Instead of having minute= and hour= keywords, now there is a
time= keyword that allows 24-hour time ranges in hh:mm-hh:mm
or hh:mm format to be specified.

Additionally, the old lists of constraint ranges are replaced
with lookup tables.  This makes everything faster and easier
to understand.  Each job still takes <2KiB of RAM, so memory
size is still very reasonable.